requirements.txt
aiogram==3.15.0
SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-dotenv==1.0.0
uvloop==0.21.0
natasha==1.6.0
razdel==0.5.0
psycopg2-binary==2.9.10